Add TestBase class
