+++ test/unit/test_modules.py
@@ -12,16 +12,19 @@
 # language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
+import contextlib
+import importlib
 import os
 import subprocess
 import sys
+import tarfile
 
-from mock import call, MagicMock, mock_open, patch
+from mock import call, mock_open, patch
 import pytest
 from six import PY2
 
 import sagemaker_containers as smc
-from test.mocks import assert_called_with, patch_context_manager, patch_with_validation
+import test
 
 builtins_open = '__builtin__.open' if PY2 else 'builtins.open'
 
@@ -84,41 +87,43 @@ def test_install_no_python_executable():
     assert str(e.value) == 'Failed to retrieve the real path for the Python executable binary'
 
 
-@patch_with_validation('importlib.import_module', smc.modules.DEFAULT_MODULE_NAME)
-@patch_with_validation('os.makedirs', '/tmp/module_dir')
-@patch_context_manager('sagemaker_containers.environment.temporary_directory', return_value='/tmp')
-@patch_with_validation('sagemaker_containers.modules.prepare', '/tmp/module_dir', smc.modules.DEFAULT_MODULE_NAME)
-@patch_with_validation('sagemaker_containers.modules.install', '/tmp/module_dir')
-@patch_with_validation('sagemaker_containers.modules.s3_download', 's3://bucket/my-module', '/tmp/tar_file')
-@patch('tarfile.open')
-def test_s3_download_import_default_name(tarfile_open):
-    smc.modules.download_and_import('s3://bucket/my-module')
-
-    with assert_called_with(tarfile_open, name='/tmp/tar_file', mode='r:gz') as t:
-        t.extractall.assert_called_with(path='/tmp/module_dir')
+@contextlib.contextmanager
+def patch_temporary_directory():
+    yield '/tmp'
 
 
-@patch_with_validation('importlib.import_module', 'another_module_name')
-@patch_with_validation('os.makedirs', '/tmp/module_dir')
-@patch_context_manager('sagemaker_containers.environment.temporary_directory', return_value='/tmp')
-@patch_with_validation('sagemaker_containers.modules.prepare', '/tmp/module_dir', 'another_module_name')
-@patch_with_validation('sagemaker_containers.modules.install', '/tmp/module_dir')
-@patch_with_validation('sagemaker_containers.modules.s3_download', 's3://bucket/my-module', '/tmp/tar_file')
-@patch('tarfile.open')
-def test_s3_download_import(tarfile_open):
-    smc.modules.download_and_import('s3://bucket/my-module', 'another_module_name')
+class TestDownloadAndImport(test.TestBase):
+    patches = [
+        patch('sagemaker_containers.environment.temporary_directory', new=patch_temporary_directory),
+        patch('sagemaker_containers.modules.prepare', autospec=True),
+        patch('sagemaker_containers.modules.install', autospec=True),
+        patch('sagemaker_containers.modules.s3_download', autospec=True),
+        patch('tarfile.open', autospec=True),
+        patch('importlib.import_module', autospec=True),
+        patch('os.makedirs', autospec=True)]
 
-    with assert_called_with(tarfile_open, name='/tmp/tar_file', mode='r:gz') as t:
-        t.extractall.assert_called_with(path='/tmp/module_dir')
+    def test_default_name(self):
+        with tarfile.open() as tar_file:
+            module = smc.modules.download_and_import('s3://bucket/my-module')
 
+            assert module == importlib.import_module(smc.modules.DEFAULT_MODULE_NAME)
 
-@patch('sagemaker_containers.modules.prepare', MagicMock(side_effect=ValueError('nothing to open')))
-@patch('sagemaker_containers.modules.s3_download', MagicMock)
-@patch('tempfile.NamedTemporaryFile', MagicMock)
-@patch_with_validation('os.makedirs', '/tmp/module_dir')
-@patch_context_manager('sagemaker_containers.environment.temporary_directory', return_value='/tmp')
-@patch(builtins_open, mock_open())
-@patch('tarfile.open', MagicMock)
-def test_s3_download_and_import_deletes_tmp_dir():
-    with pytest.raises(ValueError):
-        smc.modules.download_and_import('s3://bucket/my-module', 'another_module_name')
+            smc.modules.s3_download.assert_called_with('s3://bucket/my-module', '/tmp/tar_file')
+            os.makedirs.assert_called_with('/tmp/module_dir')
+
+            tar_file.extractall.assert_called_with(path='/tmp/module_dir')
+            smc.modules.prepare.assert_called_with('/tmp/module_dir', smc.modules.DEFAULT_MODULE_NAME)
+            smc.modules.install.assert_called_with('/tmp/module_dir')
+
+    def test_any_name(self):
+        with tarfile.open() as tar_file:
+            module = smc.modules.download_and_import('s3://bucket/my-module', 'another_module_name')
+
+            assert module == importlib.import_module('another_module_name')
+
+            smc.modules.s3_download.assert_called_with('s3://bucket/my-module', '/tmp/tar_file')
+            os.makedirs.assert_called_with('/tmp/module_dir')
+
+            tar_file.extractall.assert_called_with(path='/tmp/module_dir')
+            smc.modules.prepare.assert_called_with('/tmp/module_dir', 'another_module_name')
+            smc.modules.install.assert_called_with('/tmp/module_dir')
